Cleanup & refactor
