+++ src/Contrib.KafkaFlow.Outbox/IOutboxBackend.cs
@@ -1,8 +1,6 @@
 using Confluent.Kafka;
 
 namespace KafkaFlow.Outbox;
-
-public sealed record OutboxRecord(TopicPartition TopicPartition, Message<byte[], byte[]> Message);
 public interface IOutboxBackend
 {
     ValueTask Store(TopicPartition topicPartition, Message<byte[], byte[]> message, CancellationToken token2 = default);
