+++ src/Contrib.KafkaFlow.Outbox/IOutboxDispatcher.cs
@@ -1,0 +1,3 @@
+﻿namespace KafkaFlow.Outbox;
+
+public interface IOutboxDispatcher {}
