+++ src/Contrib.KafkaFlow.Outbox/OutboxRecord.cs
@@ -1,0 +1,5 @@
+﻿using Confluent.Kafka;
+
+namespace KafkaFlow.Outbox;
+
+public sealed record OutboxRecord(TopicPartition TopicPartition, Message<byte[], byte[]> Message);
