+++ src/Contrib.KafkaFlow.ProcessManagers.Postgres/ConfigurationBuilderExtensions.cs
@@ -1,4 +1,3 @@
-using KafkaFlow.ProcessManagers;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace KafkaFlow.ProcessManagers.Postgres;
@@ -6,5 +5,7 @@ namespace KafkaFlow.ProcessManagers.Postgres;
 public static class ConfigurationBuilderExtensions
 {
     public static IServiceCollection AddPostgresProcessManagerState(this IServiceCollection services) =>
-        services.AddSingleton<IProcessStateStore, PostgresProcessManagersStore>();
+        services
+            .AddSingleton<IProcessStateRepository, PostgresProcessStateRepository>()
+            .AddSingleton<IProcessStateStore, ProcessManagersStore>();
 }
