+++ src/Contrib.KafkaFlow.ProcessManagers/IProcessManager.cs
@@ -1,0 +1,6 @@
+﻿namespace KafkaFlow.ProcessManagers;
+
+public interface IProcessManager
+{
+    Type StateType { get; }
+}
