+++ src/Contrib.KafkaFlow.ProcessManagers/IProcessStateRepository.cs
@@ -1,0 +1,8 @@
+﻿namespace KafkaFlow.ProcessManagers;
+
+public interface IProcessStateRepository
+{
+    ValueTask<int> Persist(Type processType, string processState, Guid processId, VersionedState state);
+    ValueTask<IEnumerable<ProcessStateTableRow>> Load(Type processType, Guid processId);
+    ValueTask<int> Delete(Type processType, Guid processId, int version);
+}
