+++ src/Contrib.KafkaFlow.ProcessManagers/ProcessManager.cs
@@ -1,10 +1,5 @@
 namespace KafkaFlow.ProcessManagers;
 
-public interface IProcessManager
-{
-    Type StateType { get; }
-}
-
 public abstract class ProcessManager<TState> : IProcessManager where TState : class
 {
     public Type StateType => typeof(TState);
@@ -12,30 +7,29 @@ public abstract class ProcessManager<TState> : IProcessManager where TState : cl
 
     protected bool IsStateSet => State != null;
 
-    internal void SetState(TState? state)
-    {
-        State = state;
-    }
+    internal void SetState(TState? state) => State = state;
 
-    protected void UpdateState(TState state)
-    {
-        State = state;
-    }
+    protected void UpdateState(TState state) => State = state;
 
-    protected void FinishProcess()
-    {
-        State = null;
-    }
+    protected void FinishProcess() => State = null;
 
     protected async Task WithRequiredStateAsync(Func<TState, Task> handler)
     {
-        if (!IsStateSet) return;
+        if (!IsStateSet)
+        {
+            return;
+        }
+
         await handler(State!);
     }
 
     protected void WithRequiredState(Action<TState> handler)
     {
-        if (!IsStateSet) return;
+        if (!IsStateSet)
+        {
+            return;
+        }
+
         handler(State!);
     }
 }
