+++ src/Contrib.KafkaFlow.ProcessManagers/ProcessManagerMiddleware.cs
@@ -1,4 +1,3 @@
-using System.ComponentModel;
 using System.Transactions;
 
 namespace KafkaFlow.ProcessManagers;
