+++ src/Contrib.KafkaFlow.ProcessManagers/ProcessStateTableRow.cs
@@ -1,0 +1,3 @@
+﻿namespace KafkaFlow.ProcessManagers;
+
+public sealed record ProcessStateTableRow(string ProcessState, int Version);
