+++ tests/KafkaFlow.ProcessManagers.IntegrationTests/Fixture/IKafkaFlowFixture.cs
@@ -1,0 +1,9 @@
+﻿
+namespace KafkaFlow.ProcessManagers.IntegrationTests.Fixture;
+
+public interface IKafkaFlowFixture
+{
+    LoggingProcessStateStore ProcessStateStore { get; }
+    IMessageProducer Producer { get; }
+    string TopicName { get; }
+}
